nfse_homologacao/requirements.txt
requests>=2.31.0
pyyaml>=6.0
lxml>=4.9.3
cryptography>=41.0.0
signxml>=3.2.0
python-dateutil>=2.8.2
colorlog>=6.7.0
